folder-creation/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.6
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
setuptools==68.0.0
urllib3==2.1.0
wheel==0.41.2
